nchat: предфинальная версия приложения

+++ app/src/main/java/ru/namerpro/nchat/domain/impl/MessagesInteractorImpl.kt
@@ -1,7 +1,6 @@
 package ru.namerpro.nchat.domain.impl
 
 import com.google.gson.Gson
-import kotlinx.coroutines.ensureActive
 import ru.namerpro.nchat.commons.Constants.Companion.ENCRYPTED_FILE_PREFIX
 import ru.namerpro.nchat.commons.getFileName
 import ru.namerpro.nchat.commons.parentPath
@@ -48,17 +47,20 @@ class MessagesInteractorImpl(
         encrypter: SymmetricEncrypterContext,
         task: Task
     ): Resource<Unit> {
-        task.coroutineScope.ensureActive()
+        if (task.isCancelled) {
+            return Resource.Cancelled()
+        }
         val serverSideName = getFileName(message.devicePath)
         val filePath = "${parentPath(message.devicePath)}${ENCRYPTED_FILE_PREFIX}$serverSideName"
         RandomAccessFile(filePath, "rw").use {
             val encryptionState = encrypter.encrypt(task, message.file?.second!!, it, message.file!!.first).get()
+            if (encryptionState is EncryptionState.Cancelled) {
+                return Resource.Cancelled()
+            }
             if (encryptionState is EncryptionState.Error) {
                 return Resource.Error()
             }
-            val fileMessage = Message.Data("${serverSideName}|${message.realName}", message.date, false, message.contentType)
-            databaseInteractor.addMessage(chatId, fileMessage)
-            fileMessage.isReceived = true
+            val fileMessage = Message.Data("${serverSideName}|${message.realName}", message.date, message.time, true, message.contentType)
             val serializedMessage = toString(fileMessage, encrypter)
             val resource = messagesRepository.uploadFile(task, clientId, chatId, File(filePath), serializedMessage)
             if (resource is Resource.Error) {
@@ -66,9 +68,17 @@ class MessagesInteractorImpl(
                 Files.delete(Paths.get(message.devicePath))
                 return Resource.Error()
             }
+            fileMessage.isReceived = false
+            if (task.isCancelled) {
+                return Resource.Cancelled()
+            }
+            databaseInteractor.addMessage(chatId, fileMessage)
         }
-
-        return Resource.Success()
+        return if (task.isCancelled) {
+            Resource.Cancelled()
+        } else {
+            Resource.Success()
+        }
     }
 
     override suspend fun downloadFile(
@@ -77,23 +87,35 @@ class MessagesInteractorImpl(
         encrypter: SymmetricEncrypterContext,
         task: Task
     ): Resource<Unit> {
-        task.coroutineScope.ensureActive()
+        if (task.isCancelled) {
+            return Resource.Cancelled()
+        }
         val pathToFile = "${pathToFolder}${getFileName(message.devicePath)}"
         if (!Files.exists(Paths.get(pathToFile))) {
             val resource = messagesRepository.downloadFile(task, pathToFolder, "${ENCRYPTED_FILE_PREFIX}${getFileName(message.devicePath)}")
+            if (task.isCancelled) {
+                return Resource.Cancelled()
+            }
             if (resource is Resource.Error) {
                 return Resource.Error()
             }
             message.file = resource.data
             RandomAccessFile(pathToFile, "rw").use {
                 val state = encrypter.decrypt(task, message.file?.second!!, it, message.file!!.first).get()
+                if (task.isCancelled) {
+                    return Resource.Cancelled()
+                }
                 if (state is EncryptionState.Error) {
                     return Resource.Error()
                 }
             }
             Files.delete(Paths.get("${pathToFolder}${ENCRYPTED_FILE_PREFIX}${getFileName(message.devicePath)}"))
         }
-        return Resource.Success()
+        return if (task.isCancelled) {
+            return Resource.Cancelled()
+        } else {
+            Resource.Success()
+        }
     }
 
     private fun toString(
